refactor: consolidate utility functions into single utils.ts file
