--- src/lib/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './getLabel'
-export * from './getStyle'
-export * from './highlightMatching'
